+++ Tests/DependencyInjection/Fixtures/php/serializer_mapping.php
@@ -1,10 +1,6 @@
 <?php
 
 $container->loadFromExtension('framework', [
-    'http_method_override' => false,
-    'handle_all_throwables' => true,
-    'php_errors' => ['log' => true],
-    'annotations' => false,
     'serializer' => [
         'enable_attributes' => true,
         'mapping' => [