+++ Tests/Kernel/MicroKernelTraitTest.php
@@ -122,13 +122,7 @@ class MicroKernelTraitTest extends TestCase
 
             protected function configureContainer(ContainerConfigurator $c): void
             {
-                $c->extension('framework', [
-                    'annotations' => false,
-                    'http_method_override' => false,
-                    'handle_all_throwables' => true,
-                    'php_errors' => ['log' => true],
-                    'router' => ['utf8' => true],
-                ]);
+                $c->extension('framework', []);
                 $c->services()->set('logger', NullLogger::class);
             }